refactor(activerecord): rename and merge test files to match Rails conventions
- custom-properties.test.ts -> attributes.test.ts (38 tests)
- schema.test.ts -> active-record-schema.test.ts (14 tests)
- migration/invertible.test.ts -> invertible-migration.test.ts (29 tests)
- inverse-belongs-to + inverse-has-many + inverse-automatic -> associations/inverse-associations.test.ts (65 tests)
