--- packages/activerecord/src/associations/inverse-automatic.test.ts
@@ -1,272 +1,0 @@
-/**
- * Tests to increase Rails test coverage matching.
- * Test names are chosen to match Ruby test names from the Rails test suite.
- */
-import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
-import {
-  Base,
-  Relation,
-  Range,
-  transaction,
-  CollectionProxy,
-  association,
-  defineEnum,
-  readEnumValue,
-  RecordNotFound,
-  RecordInvalid,
-  SoleRecordExceeded,
-  ReadOnlyRecord,
-  StrictLoadingViolationError,
-  StaleObjectError,
-  columns,
-  columnNames,
-  reflectOnAssociation,
-  reflectOnAllAssociations,
-  hasSecureToken,
-  serialize,
-  registerModel,
-  composedOf,
-  acceptsNestedAttributesFor,
-  assignNestedAttributes,
-  generatesTokenFor,
-  store,
-  storedAttributes,
-  Migration,
-  Schema,
-  MigrationContext,
-  TableDefinition,
-  delegatedType,
-  enableSti,
-  registerSubclass,
-} from "../index.js";
-import {
-  Associations,
-  loadBelongsTo,
-  loadHasOne,
-  loadHasMany,
-  loadHasManyThrough,
-  processDependentAssociations,
-  updateCounterCaches,
-  setBelongsTo,
-  setHasOne,
-  setHasMany,
-} from "../associations.js";
-import {
-  OrderedOptions,
-  InheritableOptions,
-  Notifications,
-  NotificationEvent,
-} from "@rails-ts/activesupport";
-import { createTestAdapter } from "../test-adapter.js";
-import type { DatabaseAdapter } from "../adapter.js";
-import { markForDestruction, isMarkedForDestruction, isDestroyable } from "../autosave.js";
-
-// -- Helpers --
-function freshAdapter(): DatabaseAdapter {
-  return createTestAdapter();
-}
-
-describe("AutomaticInverseFindingTests", () => {
-  let adapter: DatabaseAdapter;
-  beforeEach(() => {
-    adapter = freshAdapter();
-  });
-
-  it("has one and belongs to should find inverse automatically on multiple word name", () => {
-    // Automatic inverse finding is not yet implemented; inverseOf must be explicit
-    class MixedCaseMonkey extends Base {
-      static {
-        this.attribute("name", "string");
-        this.adapter = adapter;
-      }
-    }
-    class Man extends Base {
-      static {
-        this.attribute("name", "string");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasOne.call(Man, "mixedCaseMonkey", { inverseOf: "man" });
-    Associations.belongsTo.call(MixedCaseMonkey, "man", { inverseOf: "mixedCaseMonkey" });
-    const assocs = (Man as any)._associations;
-    const hasOneAssoc = assocs.find((a: any) => a.name === "mixedCaseMonkey");
-    expect(hasOneAssoc.options.inverseOf).toBe("man");
-  });
-
-  it.skip("has many and belongs to should find inverse automatically for model in module", () => {
-    /* needs module/namespace support */
-  });
-
-  it("has one and belongs to should find inverse automatically", () => {
-    class Face extends Base {
-      static {
-        this.attribute("man_id", "integer");
-        this.adapter = adapter;
-      }
-    }
-    class Man extends Base {
-      static {
-        this.attribute("name", "string");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasOne.call(Man, "face", { inverseOf: "man" });
-    Associations.belongsTo.call(Face, "man", { inverseOf: "face" });
-    const manAssocs = (Man as any)._associations;
-    expect(manAssocs.find((a: any) => a.name === "face").options.inverseOf).toBe("man");
-  });
-
-  it("has many and belongs to should find inverse automatically", () => {
-    class Interest extends Base {
-      static {
-        this.attribute("man_id", "integer");
-        this.adapter = adapter;
-      }
-    }
-    class Man extends Base {
-      static {
-        this.attribute("name", "string");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasMany.call(Man, "interests", { inverseOf: "man" });
-    Associations.belongsTo.call(Interest, "man", { inverseOf: "interests" });
-    const manAssocs = (Man as any)._associations;
-    expect(manAssocs.find((a: any) => a.name === "interests").options.inverseOf).toBe("man");
-  });
-
-  it.skip("has many and belongs to should find inverse automatically for extension block", () => {
-    /* needs extension blocks */
-  });
-  it.skip("has many and belongs to should find inverse automatically for sti", () => {
-    /* needs STI */
-  });
-  it.skip("has one and belongs to with non default foreign key should not find inverse automatically", () => {
-    /* needs automatic inverse detection */
-  });
-  it.skip("has one and belongs to with custom association name should not find wrong inverse automatically", () => {
-    /* needs automatic inverse detection */
-  });
-  it.skip("has many and belongs to with a scope and automatic scope inversing should find inverse automatically", () => {
-    /* needs automatic scope inversing */
-  });
-  it.skip("has one and belongs to with a scope and automatic scope inversing should find inverse automatically", () => {
-    /* needs automatic scope inversing */
-  });
-  it.skip("has many with scoped belongs to does not find inverse automatically", () => {
-    /* needs automatic inverse detection */
-  });
-
-  it("has one and belongs to automatic inverse shares objects", async () => {
-    class Face extends Base {
-      static {
-        this.attribute("man_id", "integer");
-        this.attribute("description", "string");
-        this.adapter = adapter;
-      }
-    }
-    class Man extends Base {
-      static {
-        this.attribute("name", "string");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasOne.call(Man, "face", { inverseOf: "man" });
-    Associations.belongsTo.call(Face, "man", { inverseOf: "face" });
-    registerModel(Man);
-    registerModel(Face);
-    const m = await Man.create({ name: "Gordon" });
-    await Face.create({ description: "handsome", man_id: m.id });
-    const face = await loadHasOne(m, "face", { inverseOf: "man" });
-    expect(face).not.toBeNull();
-    expect((face as any)._cachedAssociations?.get("man")).toBe(m);
-  });
-
-  it("has many and belongs to automatic inverse shares objects on rating", async () => {
-    class Rating extends Base {
-      static {
-        this.attribute("score", "integer");
-        this.attribute("comment_id", "integer");
-        this.adapter = adapter;
-      }
-    }
-    class Comment extends Base {
-      static {
-        this.attribute("body", "string");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasMany.call(Comment, "ratings", { inverseOf: "comment" });
-    Associations.belongsTo.call(Rating, "comment", { inverseOf: "ratings" });
-    registerModel(Comment);
-    registerModel(Rating);
-    const c = await Comment.create({ body: "great" });
-    await Rating.create({ score: 5, comment_id: c.id });
-    const ratings = await loadHasMany(c, "ratings", { inverseOf: "comment" });
-    expect(ratings.length).toBe(1);
-    expect((ratings[0] as any)._cachedAssociations?.get("comment")).toBe(c);
-  });
-
-  it("has many and belongs to automatic inverse shares objects on comment", async () => {
-    class Comment extends Base {
-      static {
-        this.attribute("body", "string");
-        this.attribute("post_id", "integer");
-        this.adapter = adapter;
-      }
-    }
-    class Post extends Base {
-      static {
-        this.attribute("title", "string");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasMany.call(Post, "comments", { inverseOf: "post" });
-    Associations.belongsTo.call(Comment, "post", { inverseOf: "comments" });
-    registerModel(Post);
-    registerModel(Comment);
-    const p = await Post.create({ title: "hello" });
-    await Comment.create({ body: "nice", post_id: p.id });
-    const comments = await loadHasMany(p, "comments", { inverseOf: "post" });
-    expect(comments.length).toBe(1);
-    expect((comments[0] as any)._cachedAssociations?.get("post")).toBe(p);
-  });
-
-  it("belongs to should find inverse has many automatically", async () => {
-    class Interest extends Base {
-      static {
-        this.attribute("topic", "string");
-        this.attribute("man_id", "integer");
-        this.adapter = adapter;
-      }
-    }
-    class Man extends Base {
-      static {
-        this.attribute("name", "string");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasMany.call(Man, "interests", { inverseOf: "man" });
-    Associations.belongsTo.call(Interest, "man", { inverseOf: "interests" });
-    registerModel(Man);
-    registerModel(Interest);
-    const m = await Man.create({ name: "Gordon" });
-    const i = await Interest.create({ topic: "stamps", man_id: m.id });
-    const parent = await loadBelongsTo(i, "man", { inverseOf: "interests" });
-    expect(parent).not.toBeNull();
-    expect((parent as any)._cachedAssociations?.get("interests")).toBe(i);
-  });
-
-  it.skip("polymorphic and has many through relationships should not have inverses", () => {
-    /* needs automatic inverse detection */
-  });
-  it.skip("polymorphic has one should find inverse automatically", () => {
-    /* needs automatic inverse detection for polymorphic */
-  });
-  it.skip("has many inverse of derived automatically despite of composite foreign key", () => {
-    /* needs composite FK */
-  });
-  it.skip("belongs to inverse of derived automatically despite of composite foreign key", () => {
-    /* needs composite FK */
-  });
-});
