--- packages/activerecord/src/associations/inverse-belongs-to.test.ts
@@ -1,254 +1,0 @@
-/**
- * Tests to increase Rails test coverage matching.
- * Test names are chosen to match Ruby test names from the Rails test suite.
- */
-import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
-import {
-  Base,
-  Relation,
-  Range,
-  transaction,
-  CollectionProxy,
-  association,
-  defineEnum,
-  readEnumValue,
-  RecordNotFound,
-  RecordInvalid,
-  SoleRecordExceeded,
-  ReadOnlyRecord,
-  StrictLoadingViolationError,
-  StaleObjectError,
-  columns,
-  columnNames,
-  reflectOnAssociation,
-  reflectOnAllAssociations,
-  hasSecureToken,
-  serialize,
-  registerModel,
-  composedOf,
-  acceptsNestedAttributesFor,
-  assignNestedAttributes,
-  generatesTokenFor,
-  store,
-  storedAttributes,
-  Migration,
-  Schema,
-  MigrationContext,
-  TableDefinition,
-  delegatedType,
-  enableSti,
-  registerSubclass,
-} from "../index.js";
-import {
-  Associations,
-  loadBelongsTo,
-  loadHasOne,
-  loadHasMany,
-  loadHasManyThrough,
-  processDependentAssociations,
-  updateCounterCaches,
-  setBelongsTo,
-  setHasOne,
-  setHasMany,
-} from "../associations.js";
-import {
-  OrderedOptions,
-  InheritableOptions,
-  Notifications,
-  NotificationEvent,
-} from "@rails-ts/activesupport";
-import { createTestAdapter } from "../test-adapter.js";
-import type { DatabaseAdapter } from "../adapter.js";
-import { markForDestruction, isMarkedForDestruction, isDestroyable } from "../autosave.js";
-
-// -- Helpers --
-function freshAdapter(): DatabaseAdapter {
-  return createTestAdapter();
-}
-
-describe("InverseBelongsToTests", () => {
-  let adapter: DatabaseAdapter;
-  beforeEach(() => {
-    adapter = freshAdapter();
-  });
-
-  function makeModels() {
-    class Man extends Base {
-      static {
-        this.attribute("name", "string");
-        this.adapter = adapter;
-      }
-    }
-    class Face extends Base {
-      static {
-        this.attribute("description", "string");
-        this.attribute("man_id", "integer");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasOne.call(Man, "face", { inverseOf: "man" });
-    Associations.belongsTo.call(Face, "man", { inverseOf: "face" });
-    registerModel(Man);
-    registerModel(Face);
-    return { Man, Face };
-  }
-
-  it("child instance should be shared with parent on find", async () => {
-    const { Man, Face } = makeModels();
-    const m = await Man.create({ name: "Gordon" });
-    const f = await Face.create({ description: "pretty", man_id: m.id });
-    const parent = await loadBelongsTo(f, "man", { inverseOf: "face" });
-    expect(parent).not.toBeNull();
-    expect((parent as any)._cachedAssociations?.get("face")).toBe(f);
-  });
-
-  it("eager loaded child instance should be shared with parent on find", async () => {
-    const { Man, Face } = makeModels();
-    const m = await Man.create({ name: "Gordon" });
-    await Face.create({ description: "pretty", man_id: m.id });
-    const faces = await Face.all().includes("man").toArray();
-    expect(faces.length).toBe(1);
-    const parent = (faces[0] as any)._preloadedAssociations?.get("man");
-    expect(parent).not.toBeNull();
-    expect((parent as any)._cachedAssociations?.get("face")).toBe(faces[0]);
-  });
-
-  it("child instance should be shared with newly built parent", () => {
-    const { Man, Face } = makeModels();
-    const f = new Face({ description: "pretty" });
-    const m = new Man({ name: "Gordon" });
-    // Manually set inverse
-    (f as any)._cachedAssociations = new Map();
-    (f as any)._cachedAssociations.set("man", m);
-    expect((f as any)._cachedAssociations.get("man")).toBe(m);
-  });
-
-  it("child instance should be shared with newly created parent", async () => {
-    const { Man, Face } = makeModels();
-    const m = await Man.create({ name: "Gordon" });
-    const f = await Face.create({ description: "pretty", man_id: m.id });
-    const parent = await loadBelongsTo(f, "man", { inverseOf: "face" });
-    expect(parent).not.toBeNull();
-    expect((parent as any)._cachedAssociations?.get("face")).toBe(f);
-  });
-
-  it("with has many inversing should try to set inverse instances when the inverse is a has many", async () => {
-    class Author extends Base {
-      static {
-        this.attribute("name", "string");
-        this.adapter = adapter;
-      }
-    }
-    class Book extends Base {
-      static {
-        this.attribute("title", "string");
-        this.attribute("author_id", "integer");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasMany.call(Author, "books", { inverseOf: "author" });
-    Associations.belongsTo.call(Book, "author", { inverseOf: "books" });
-    registerModel(Author);
-    registerModel(Book);
-    const a = await Author.create({ name: "Alice" });
-    const b = await Book.create({ title: "Wonderland", author_id: a.id });
-    const parent = await loadBelongsTo(b, "author", { inverseOf: "books" });
-    expect(parent).not.toBeNull();
-    expect((parent as any)._cachedAssociations?.get("books")).toBe(b);
-  });
-
-  it.skip("with has many inversing should have single record when setting record through attribute in build method", () => {
-    /* needs has_many inversing push */
-  });
-  it.skip("with has many inversing does not trigger association callbacks on set when the inverse is a has many", () => {
-    /* needs callback tracking */
-  });
-  it.skip("with has many inversing does not add duplicate associated objects", () => {
-    /* needs has_many inversing */
-  });
-  it.skip("with has many inversing does not add unsaved duplicate records when collection is loaded", () => {
-    /* needs collection tracking */
-  });
-  it.skip("with has many inversing does not add saved duplicate records when collection is loaded", () => {
-    /* needs collection tracking */
-  });
-
-  it("recursive model has many inversing", async () => {
-    class Node extends Base {
-      static {
-        this.attribute("name", "string");
-        this.attribute("node_id", "integer");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasMany.call(Node, "children", {
-      className: "Node",
-      foreignKey: "node_id",
-      inverseOf: "parent",
-    });
-    Associations.belongsTo.call(Node, "parent", {
-      className: "Node",
-      foreignKey: "node_id",
-      inverseOf: "children",
-    });
-    registerModel(Node);
-    const parent = await Node.create({ name: "root" });
-    const child = await Node.create({ name: "leaf", node_id: parent.id });
-    const foundParent = await loadBelongsTo(child, "parent", {
-      className: "Node",
-      foreignKey: "node_id",
-      inverseOf: "children",
-    });
-    expect(foundParent).not.toBeNull();
-    expect((foundParent as any)._cachedAssociations?.get("children")).toBe(child);
-  });
-
-  it.skip("recursive inverse on recursive model has many inversing", () => {
-    /* needs deep recursive inverse */
-  });
-  it.skip("unscope does not set inverse when incorrect", () => {
-    /* needs unscope support */
-  });
-  it.skip("or does not set inverse when incorrect", () => {
-    /* needs or-query inverse checking */
-  });
-  it("child instance should be shared with replaced via accessor parent", async () => {
-    const { Man, Face } = makeModels();
-    const m1 = await Man.create({ name: "Gordon" });
-    const f = await Face.create({ description: "pretty", man_id: m1.id });
-    const m2 = await Man.create({ name: "New Guy" });
-    setBelongsTo(f, "man", m2, { inverseOf: "face" });
-    expect((f as any)._cachedAssociations.get("man")).toBe(m2);
-    expect((m2 as any)._cachedAssociations?.get("face")).toBe(f);
-  });
-  it.skip("trying to use inverses that dont exist should raise an error", () => {
-    /* needs inverse validation */
-  });
-  it.skip("trying to use inverses that dont exist should have suggestions for fix", () => {
-    /* needs inverse validation */
-  });
-
-  it("building has many parent association inverses one record", async () => {
-    class Author extends Base {
-      static {
-        this.attribute("name", "string");
-        this.adapter = adapter;
-      }
-    }
-    class Book extends Base {
-      static {
-        this.attribute("title", "string");
-        this.attribute("author_id", "integer");
-        this.adapter = adapter;
-      }
-    }
-    Associations.hasMany.call(Author, "books", { inverseOf: "author" });
-    Associations.belongsTo.call(Book, "author", { inverseOf: "books" });
-    registerModel(Author);
-    registerModel(Book);
-    const a = await Author.create({ name: "Alice" });
-    const proxy = association(a, "books");
-    const b = proxy.build({ title: "New Book" });
-    expect(b.readAttribute("author_id")).toBe(a.id);
-  });
-});
